+++ my_caesar_cli.js
@@ -3,7 +3,7 @@ const program = new commander.Command();
 const fs = require('fs');
 const checkExistence = require('./check-existence');
 const CaesarTransformer = require('./CaesarTransformer');
-const validateArguments = require('./validateArguments');
+const validateArguments = require('./validate-arguments');
 
 program
   .requiredOption('-s, --shift <value>', 'a shift')
